Extract QuizMeta and QuizQuestion into frontend /lib
